Split out the deterministic output into it's own function to generate the fixed text output
Adjusted demo.py so that it now calls this function

Had to add requests python library in requirements.txt

Created telegram_client so that
also added telegram chat information in the environment variables so that the client is able to send the message to your conversation

Created the fixed_message script which reads in the current days sleep and previous 7 days from the database. calls the deterministic_output.py and then send the output via the telegram client.