Created last_sleep_sent.json to save the date of the last sleep that was sent, so we know when to trigger a telegram send on a new sleep being uploaded to our database as well as avoid doing duplicate sends.
In fixed_message.py added the two functions load_last_sent_key() and save_last_sent_key() and implemented them in the main function to check if the last key matches the previous days and exits otherwise if it doesn't match it is a new day and send the message and loads the new key into the file.

Did not commit the sleep state .json file as the directory is automatically created from the fixed_message.py function. It is better this way because editing the file is locked because it was created in the container rather than by the user. This makes it less likely that the user will accidently corrupt the file.