Created a seperate container in compose.addon.yml that runs the scheduler to call the fixed_message at a fixed interval defined as CHECK_INTERVAL_SECONDS. Furthermore the fixed_message is called is provided in this script so to update it all you need to do is change the function that is called here.
The sleep-checkins-scheduler container calls the scheduler.py function. The schedular function sleeps for the given period which is the fixed interval as well as some jitter sections to randomize the specific call time. The jitter time is also defined in compose.addon.yml

We updated the fixed message so the main function is essentially renamed to run_once() so that we can call this from the schedular and we return a T/F dpending on if a message is sent