Ok we implemented recieving messages via telegram a new container telegram-listener was added that runs until stopped.
The telegram_listener.py is the most crucial function that uses a longPoll via the get updates to look for new chats. It saves the state of the chat of the last message in the data folder under telegram_listener_state.json. It does some basic santization of the input removing hidden characters. However, one thing that may be a limitation is that we limit the size of the input to be 512 characters in the sanitize_text() function, so depending on how we use the app this may limit us and may need to be removed or handled differently in the database.

If telegram_listener.py recognizes the message is new than we call the write_telegram_journal_entry() which is contained in journal_store.py along with a helper function to connect the influx database. It is written into the database as a SleepJournal Measurement.

We also updated the sleep_data_export.py for this new measurement so that it exports a .jsonl as well as a .csv file of these journal entries that have been recieved form the telegram chat. This can be used for exporting or for backing up these insights.

Also, we created the delete_sleepjournal_entries.py which removes all the SleepJournal measurements from the influx database, to clean things up.

Also updated deterministic_output.py so that there is a line at the end prompting you for a journal input.

Finally we wrote a little bash script so when changes are made to scripts or we are starting the program for the first time on a new machine called rebuild.sh it also provides the helpful output for one-shotting with a container to do an export, single message (need to delete existing key in data folder if already sent, to defeat dedupe) or to delete all the current journal entry measurements.