Corrected fixed_message so it is only asking me once at the end not after each line.
updated the read-me to create a better summary of the current project as well as setup a basic quickstart guide of the key steps. And detailing the setup on the telegram bot for messaging.

Updated the .env file to take in MAX_JOURNAL_CHARACTERS as a configurable limit and updated the respective telegram_listener.py santize input function. Also implemented error handling if what is recieved is not a number.